requirements.txt
Flask==1.1.2
requests==2.25.1
psycopg2-binary==2.8.6
gunicorn==20.0.4
google-cloud-storage==1.36.0
